Implement Common ItemControl
- Create Test Harness
- Integrate into Library
- Integrate into Navigator

Related work items: #785, #784, #786, #787, #788

+++ src/Chem4Word.V3/UI/WPF/Chem4WordSettingsControl.xaml.cs
@@ -20,7 +20,6 @@ using Chem4Word.Core;
 using Chem4Word.Core.Helpers;
 using Chem4Word.Core.UI.Forms;
 using Chem4Word.Core.UI.Wpf;
-using Chem4Word.Libraries;
 using Chem4Word.Libraries.Database;
 using Chem4Word.Model2;
 using Chem4Word.Model2.Converters.CML;
@@ -385,18 +384,16 @@ namespace Chem4Word.UI.WPF
                         sb.AppendLine("Do you want to import the Gallery structures into the Library?");
                         //sb.AppendLine("(This cannot be undone.)")
                         dr = UserInteractions.AskUserYesNo(sb.ToString());
-                        if (dr == Forms.DialogResult.Yes)
+                        if (dr == Forms.DialogResult.Yes
+                            && File.Exists(doneFile))
                         {
-                            if (File.Exists(doneFile))
+                            sb = new StringBuilder();
+                            sb.AppendLine($"Files have been imported already from '{selectedFolder}'");
+                            sb.AppendLine("Do you want to rerun the import?");
+                            dr = UserInteractions.AskUserYesNo(sb.ToString());
+                            if (dr == Forms.DialogResult.Yes)
                             {
-                                sb = new StringBuilder();
-                                sb.AppendLine($"Files have been imported already from '{selectedFolder}'");
-                                sb.AppendLine("Do you want to rerun the import?");
-                                dr = UserInteractions.AskUserYesNo(sb.ToString());
-                                if (dr == Forms.DialogResult.Yes)
-                                {
-                                    File.Delete(doneFile);
-                                }
+                                File.Delete(doneFile);
                             }
                         }
 
@@ -404,15 +401,7 @@ namespace Chem4Word.UI.WPF
                         {
                             int fileCount = 0;
 
-                            var lib = new Libraries.Database.Library(Globals.Chem4WordV3.Telemetry,
-                                                                     new LibrarySettings
-                                                                     {
-                                                                         ParentTopLeft = Globals.Chem4WordV3.WordTopLeft,
-                                                                         ProgramDataPath = Globals.Chem4WordV3.AddInInfo.ProgramDataPath,
-                                                                         PreferredBondLength = Globals.Chem4WordV3.SystemOptions.BondLength,
-                                                                         SetBondLengthOnImport = Globals.Chem4WordV3.SystemOptions.SetBondLengthOnImportFromLibrary,
-                                                                         RemoveExplicitHydrogensOnImport = Globals.Chem4WordV3.SystemOptions.RemoveExplicitHydrogensOnImportFromLibrary
-                                                                     });
+                            var lib = new Libraries.Database.Library(Globals.Chem4WordV3.Telemetry, Globals.Chem4WordV3.LibraryOptions);
                             var transaction = lib.StartTransaction();
 
                             try
@@ -512,20 +501,12 @@ namespace Chem4Word.UI.WPF
                         }
                         if (doExport == Forms.DialogResult.Yes)
                         {
-                            Libraries.Database.Library lib = new Libraries.Database.Library(Globals.Chem4WordV3.Telemetry,
-                                                                                            new LibrarySettings
-                                                                                            {
-                                                                                                ParentTopLeft = Globals.Chem4WordV3.WordTopLeft,
-                                                                                                ProgramDataPath = Globals.Chem4WordV3.AddInInfo.ProgramDataPath,
-                                                                                                PreferredBondLength = Globals.Chem4WordV3.SystemOptions.BondLength,
-                                                                                                SetBondLengthOnImport = Globals.Chem4WordV3.SystemOptions.SetBondLengthOnImportFromLibrary,
-                                                                                                RemoveExplicitHydrogensOnImport = Globals.Chem4WordV3.SystemOptions.RemoveExplicitHydrogensOnImportFromLibrary
-                                                                                            });
+                            Libraries.Database.Library lib = new Libraries.Database.Library(Globals.Chem4WordV3.Telemetry, Globals.Chem4WordV3.LibraryOptions);
 
                             int exported = 0;
                             int progress = 0;
 
-                            List<ChemistryDTO> dto = lib.GetAllChemistry(null);
+                            List<ChemistryDTO> dto = lib.GetAllChemistry();
                             int total = dto.Count;
                             if (total > 0)
                             {
@@ -598,15 +579,7 @@ namespace Chem4Word.UI.WPF
                     UserInteractions.AskUserYesNo(sb.ToString(), Forms.MessageBoxDefaultButton.Button2);
                 if (dr == Forms.DialogResult.Yes)
                 {
-                    var lib = new Libraries.Database.Library(Globals.Chem4WordV3.Telemetry,
-                                                             new LibrarySettings
-                                                             {
-                                                                 ParentTopLeft = Globals.Chem4WordV3.WordTopLeft,
-                                                                 ProgramDataPath = Globals.Chem4WordV3.AddInInfo.ProgramDataPath,
-                                                                 PreferredBondLength = Globals.Chem4WordV3.SystemOptions.BondLength,
-                                                                 SetBondLengthOnImport = Globals.Chem4WordV3.SystemOptions.SetBondLengthOnImportFromLibrary,
-                                                                 RemoveExplicitHydrogensOnImport = Globals.Chem4WordV3.SystemOptions.RemoveExplicitHydrogensOnImportFromLibrary
-                                                             });
+                    var lib = new Libraries.Database.Library(Globals.Chem4WordV3.Telemetry, Globals.Chem4WordV3.LibraryOptions);
 
                     lib.DeleteAllChemistry();
                     Globals.Chem4WordV3.CloseLibrary();
